+++ src/sequential/random.h
@@ -15,7 +15,7 @@ public:
     virtual ~Random();
     int get_initial_city(int n_cities);
     int get_next_city(Ant *ant, int start, int n_cities, Dataloader *dataloader,
-                      double *phero, double alpha, double beta);
+                      float *phero, double alpha, double beta);
 };
